Fix GCC >= 14 compilation failures

+++ daphne/src/thread/generic/SDL_systhread_c.h
@@ -21,6 +21,6 @@
 #include "../../SDL_internal.h"
 
 /* Stub until we implement threads on this platform */
-typedef int SYS_ThreadHandle;
+typedef unsigned long int SYS_ThreadHandle;
 
 /* vi: set ts=4 sw=4 expandtab: */
